+++ src/main/kotlin/com/sendy/application/usecase/account/command/CreateAccountService.kt
@@ -2,46 +2,15 @@ package com.sendy.application.usecase.account.command
 
 import com.sendy.application.dto.account.CreateAccountRequest
 import com.sendy.application.dto.account.CreateAccountResponse
-import com.sendy.domain.account.AccountEntity
-import com.sendy.domain.account.AccountRepository
-import com.sendy.domain.account.AccountStatus
-import com.sendy.support.util.getTsid
-import com.sendy.support.util.AccountNumberValidator
-import com.sendy.support.util.Aes256Util
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.time.LocalDateTime
 
 @Transactional
 @Service
 class CreateAccountService(
-    private val accountRepository: AccountRepository,
-    private val createAccountNumberUseCase: CreateAccountNumberUseCase,
-    private val Aes256Util: Aes256Util
+    private val createAccountUseCase: CreateAccountUseCaseImpl
 ) : CreateAccountUseCase {
     override fun execute(request: CreateAccountRequest): CreateAccountResponse {
-        // 서버에서 계좌번호를 자동 생성
-        val generatedAccountNumber = createAccountNumberUseCase.generate()
-        
-        // 생성된 계좌번호 형식 검증 (321-xxxx-xx-xxxx 형식)
-        AccountNumberValidator.validateFormat(generatedAccountNumber)
-        val encryptedPassword = Aes256Util.encrypt(request.password)
-        return accountRepository
-            .save(
-                AccountEntity(
-                    id = getTsid(),
-                    accountNumber = generatedAccountNumber, // 서버에서 생성한 계좌번호 사용
-                    userId = request.userId,
-                    password = encryptedPassword, // 클라이언트에서 받은 비밀번호
-                    status = AccountStatus.ACTIVE,
-                    isPrimary = request.isPrimary,
-                    isLimitedAccount = request.isLimitedAccount,
-                    createdAt = LocalDateTime.now(),
-                    updatedAt = LocalDateTime.now(),
-                    balance = 0L,
-                ),
-            ).let {
-                CreateAccountResponse(it.accountNumber)
-            }
+        return createAccountUseCase.execute(request)
     }
 }
